requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
filelock>=3.12.0
